Improvements to client setup
