+++ test/cli.test.js
@@ -1,0 +1,3 @@
+const test = require('tape')
+
+test('cli') // needs tests
