+++ test/server.test.js
@@ -1,0 +1,3 @@
+const test = require('tape')
+
+test('server') // needs tests
